[REALESTATE-21] 수량 long 변경

+++ src/main/java/com/labshigh/realestate/internal/api/marketItem/model/response/MarketItemDetailResponseModel.java
@@ -22,8 +22,8 @@ public class MarketItemDetailResponseModel {
   private Boolean usedFlag;
 
   private long itemUid;
-  private int quantity;
-  private int currentQuantity;
+  private long quantity;
+  private long currentQuantity;
   @JsonFormat(pattern = Constants.JSONFY_DATE_FORMAT)
   private LocalDateTime startAt;
   @JsonFormat(pattern = Constants.JSONFY_DATE_FORMAT)
