+++ src/main/java/com/labshigh/realestate/internal/api/marketItem/model/response/MarketItemListResponseModel.java
@@ -21,8 +21,8 @@ public class MarketItemListResponseModel {
   private Boolean usedFlag;
 
   private long itemUid;
-  private int quantity;
-  private int currentQuantity;
+  private long quantity;
+  private long currentQuantity;
   @JsonFormat(pattern = Constants.JSONFY_DATE_FORMAT)
   private LocalDateTime startAt;
   @JsonFormat(pattern = Constants.JSONFY_DATE_FORMAT)
